// src/app/Authentication/models/IProfileRequest.Model.ts
export interface IProfileRequest {
    personalImage: string;
    firstName: string;
    lastName: string;
    fullNameArabic: string;
    fullNameEnglish: string;
    dateOfBirth: string;
    nationality: string;
    gender: number;
    country: string;
    city: string;
    fullAddress: string;
    aboutMe: string;
}